headhunter/requirements.txt
alembic==1.6.5
appdirs==1.4.4
astroid==2.6.5
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
flake8==3.9.2
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.6
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
regex==2021.7.6
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
tomli==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
